地形の当たり判定、簡易的な昼夜表現実装 (#6)
* feat : 地形との当たり判定を入れたシーン追加

feat : モデル追加

feat : TransformにImGui用の参照返し版を追加

feat : パーリンノイズ用.h/cpp追加

feat : 乱数生成器追加

feat : 障害物クラス追加

* fix : GameObjectにDestroy()追加

* fix : Updateのデルタタイムをuint64からfloatに変更

feat : 簡易的な昼夜表現実装

feat : ゲーム内の時間とタイムスケールを管理するTimeクラスの実装

* fix : Enemyのプレイヤーnullptrチェックを改善

+++ 3D_TPS/system/Billboard/Billboard.h
@@ -25,7 +25,7 @@ public:
 	void Init(void) override;
 	void Init(int width, int height, const std::string& texfilename, FreeCamera* cam);
 
-	void Update(const uint64_t deltatime) override;
+	void Update(const float deltatime) override;
 
 	void Draw(void) const override;
